[Refactor][Kurt] - Moved students route under protected routes folder
