+++ src/app/(private-routes)/student/account/page.js
@@ -1,0 +1,5 @@
+const StudentAcountPage = () => {
+    return <div></div>;
+};
+
+export default StudentAcountPage;
