+++ src/app/(private-routes)/student/enrollment/page.js
@@ -1,0 +1,5 @@
+const StudentEnrollmentPage = () => {
+    return <div></div>;
+};
+
+export default StudentEnrollmentPage;
