+++ src/app/(private-routes)/student/grades/page.js
@@ -1,0 +1,6 @@
+
+const StudentGradesPage = () => {
+    return <div></div>;
+};
+
+export default StudentGradesPage;
