+++ src/app/(private-routes)/student/information/page.js
@@ -1,0 +1,5 @@
+const StudentInformationPage = () => {
+    return <div></div>;
+};
+
+export default StudentInformationPage;
