+++ src/app/(private-routes)/student/page.js
@@ -1,0 +1,5 @@
+const StudentLoginPage = () => {
+    return <>test</>;
+};
+
+export default StudentLoginPage;
