--- src/app/student/page.js
@@ -1,13 +1,0 @@
-import React from 'react';
-
-import StudentLogin from './student-login';
-
-const StudentLoginPage = () => {
-    return (
-        <>
-            <StudentLogin />
-        </>
-    );
-};
-
-export default StudentLoginPage;
